requirements.txt
beautifulsoup4==4.9.3
boto3==1.17.90
botocore==1.20.90
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dash==0.39.0
dash-core-components==0.44.0
dash-daq==0.1.0
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
Flask==2.0.1
Flask-Compress==1.10.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
numpy==1.21.0
paho-mqtt==1.5.1
pandas==1.2.5
Pillow==8.2.0
plotly==5.0.0
psycopg2==2.9.1
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.3.1
tenacity==7.0.0
urllib3==1.26.5
Werkzeug==2.0.1